backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
